server/pyproject.toml
[project]
name = "server"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "lark>=1.2.2",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "z3-solver>=4.15.3.0",
]

[dependency-groups]
dev = [
    "mongomock>=4.3.0",
    "mongomock-motor>=0.0.36",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
